+++ packages/core/index.d.ts
@@ -6,6 +6,7 @@ export {
   MessageMeta,
   Identifier
 } from './channel/index.js'
+export { ChannelsCache } from './channels-cache/index.js'
 export { Transport } from './transport/index.js'
 export { Encoder, JSONEncoder } from './encoder/index.js'
 export {
@@ -20,13 +21,15 @@ export {
 } from './protocol/index.js'
 export { Hub } from './hub/index.js'
 export { Logger, LogLevel, BaseLogger, NoopLogger } from './logger/index.js'
+export { stringifyParams } from './stringify-params/index.js'
 
 export { ActionCableProtocol } from './action_cable/index.js'
 export {
   CableOptions,
   Cable,
   NoConnectionError,
-  CableEvents
+  CableEvents,
+  Cache
 } from './cable/index.js'
 export {
   Monitor,