+++ packages/core/index.js
@@ -6,12 +6,14 @@ export {
   CommandError,
   StaleConnectionError
 } from './protocol/index.js'
+export { ChannelsCache } from './channels-cache/index.js'
 export { Hub } from './hub/index.js'
 export { BaseLogger, NoopLogger } from './logger/index.js'
 export { JSONEncoder } from './encoder/index.js'
 export { ActionCableProtocol } from './action_cable/index.js'
 export { Cable, NoConnectionError } from './cable/index.js'
 export { Monitor, backoffWithJitter } from './monitor/index.js'
+export { stringifyParams } from './stringify-params/index.js'
 export { WebSocketTransport } from './websocket/index.js'
 export {
   createCable,