+++ packages/core/stringify-params/index.d.ts
@@ -1,0 +1,3 @@
+type Params = { [token: string]: boolean | number | string }
+
+export function stringifyParams(params: Params | void | null): string